refactor: создает отдельный модуль для эффектов
