+++ js/upload-picture.js
@@ -1,6 +1,6 @@
 import { addScaleListeners } from './upload-picture-scale.js';
-import { validateUploadPictureForm } from './upload-picture-validation.js';
 import { createSlider, setupSlider, destroySlider } from './upload-picture-slider.js';
+import { addValidators, isValidForm, resetPristine, addEffectListener } from './upload-picture-validation.js';
 import { showMessage, createSuccessMessage, createErrorMessage } from './upload-picture-fetch-messages.js';
 import { sendData } from './server-data.js';
 
@@ -10,15 +10,12 @@ const pictureUploadPreview = document.querySelector('.img-upload__preview img');
 const pictureEdit = document.querySelector('.img-upload__overlay');
 const submitButton = document.querySelector('.img-upload__submit');
 const closeButton = document.querySelector('.img-upload__cancel');
-const pictureScaleInput = document.querySelector('.scale__control--value');
-
-const effectsList = document.querySelector('.effects__list');
 const checkedEffectInput = document.querySelector('.effects__radio[checked]');
 
 const onUploadPictureFormSubmit = (evt) => {
   evt.preventDefault();
 
-  if (validateUploadPictureForm()) {
+  if (isValidForm()) {
     blockSubmitButton();
     sendData(new FormData(evt.target))
       .then(() => {
@@ -34,14 +31,6 @@ const onUploadPictureFormSubmit = (evt) => {
   }
 };
 
-const onEffectInputClick = (evt) => {
-  if (evt.target.closest('.effects__radio')) {
-    pictureUploadPreview.className = '';
-    pictureUploadPreview.classList.add(`effects__preview--${evt.target.value}`);
-    setupSlider(evt.target.value);
-  }
-};
-
 const onCloseButtonClick = () => closePictureUpload();
 
 const onDocumentKeydown = (evt) => {
@@ -52,7 +41,6 @@ const onDocumentKeydown = (evt) => {
 
 const addFormBaseListeners = () => {
   pictureUploadForm.addEventListener('submit', onUploadPictureFormSubmit);
-  effectsList.addEventListener('change', onEffectInputClick);
   closeButton.addEventListener('click', onCloseButtonClick);
   document.addEventListener('keydown', onDocumentKeydown);
 };
@@ -70,8 +58,6 @@ function unblockSubmitButton() {
 function defaultSetupPictureUpload() {
   pictureUploadForm.reset();
   pictureUploadPreview.style = null;
-  pictureUploadPreview.className = 'effects__preview--none';
-  pictureScaleInput.defaultValue = '100%';
 }
 
 function openPictureUpload() {
@@ -83,6 +69,7 @@ function openPictureUpload() {
 
 function closePictureUpload() {
   destroySlider();
+  resetPristine();
   defaultSetupPictureUpload();
   document.body.classList.remove('modal-open');
   pictureEdit.classList.add('hidden');
@@ -92,6 +79,8 @@ const initUploadPictureModule = () => {
   pictureUploadInput.addEventListener('change', openPictureUpload);
   addFormBaseListeners();
   addScaleListeners();
+  addEffectListener();
+  addValidators();
 };
 
 export { initUploadPictureModule };
